// libs/core/src/lib/enums/transaction-state.ts
export enum TransactionState {
  STATE = 'STATE',
  SUBSTATE = 'SUBSTATE',
  OUTCOME = 'OUTCOME',
  TEXTATTRIBUTE6 = 'textAttribute6',
  UPDATEDDATE = 'UPDATEDDATE',
  ASSIGNEDDATE = 'ASSIGNEDDATE',
  SCA_PARTITION_ID = 'SCA_PARTITION_ID',
  ASSIGNEES = 'assignees',
  ISGROUP = 'ISGROUP',
  VERSIONREASON = 'VERSIONREASON',
  PROTECTEDTEXTATTRIBUTE1 = 'PROTECTEDTEXTATTRIBUTE1',
  PROTECTEDTEXTATTRIBUTE7 = 'PROTECTEDTEXTATTRIBUTE7'
}