// libs/core/src/lib/enums/search-column.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

/* Legal entities enum */
export enum SearchColumn {
  TITILEARABIC = 'textAttribute4',
  TITLEENGLISH = 'textAttribute3',
  TRANSACTIONID = 'textAttribute2',
  OLABREACHED = 'PROTECTEDTEXTATTRIBUTE7',
  SWIMLANEROLE='swimlaneRole'
}